requirements.txt
streamlit
pymupdf
faiss-cpu
sentence-transformers
numpy
pandas
python-dotenv
requests
tiktoken